+++ src/environment/bus_world/bus_world_events/new_bus.rs
@@ -20,12 +20,12 @@ impl NewBusesJson {
 
 pub struct NewBusEvent {
     uid: usize,
-    timestamp: f64,
+    timestamp: usize,
     data: String,
 }
 
 impl NewBusEvent {
-    pub fn new(uid: usize, timestamp: f64, data: String) -> NewBusEvent {
+    pub fn new(uid: usize, timestamp: usize, data: String) -> NewBusEvent {
         NewBusEvent {
             uid,
             timestamp,
@@ -43,7 +43,7 @@ impl Event for NewBusEvent {
         self.uid
     }
 
-    fn get_time_stamp(&self) -> f64 {
+    fn get_time_stamp(&self) -> usize {
         self.timestamp
     }
 
@@ -69,10 +69,10 @@ mod tests {
 
     #[test]
     fn create_bus_event() {
-        let bus_event = NewBusEvent::new(1, 0.0, String::from("Hello world!"));
+        let bus_event = NewBusEvent::new(1, 0, String::from("Hello world!"));
         assert_eq!(bus_event.get_event_type(), "NewBus");
         assert_eq!(bus_event.get_uid(), 1);
-        assert_eq!(bus_event.get_time_stamp(), 0.0);
+        assert_eq!(bus_event.get_time_stamp(), 0);
         assert_eq!(bus_event.get_data().unwrap(), "Hello world!");
     }
 }
